Squashed commit of the following:
commit 441bb67a081d5b997d81595d6dfdc0c29a9a7e19
Author: Sam Atkins <sam@trufflesec.com>
Date:   Mon Jan 16 09:45:34 2023 -0700

    Migration to make tab content nullable

commit b5f8c93ec57dfab18944e200f5a2cdcd2b3d6cef
Author: atkins <sam@trufflesec.com>
Date:   Sat Jan 7 21:29:49 2023 -0700

    Fix noodle issues and add songbook title to add (#48)

commit 4bd1ff944a223ddc969be83058dc5e7c74ecf946
Author: atkins <sam@trufflesec.com>
Date:   Sat Jan 7 20:52:49 2023 -0700

    Security updates (#47)

commit f334abe7a5dc68ac49d282eeb1da2310d6f20708
Author: Jack Conway <52581258+jackconwayiv@users.noreply.github.com>
Date:   Fri Jan 6 09:26:03 2023 -0700

    Componentize the hamburger menu. (#46)
