+++ frontend/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "livepowerhour.com",
+  "name": "songbook generator",
   "icons": [
     {
       "src": "favicon.ico",
